Fix download directory in default configs
Commit a07d6b9d56955bfe0165b5f4ea6d3ed7a92ecff6 changed the meaning
of $(BASE_DIR). Previously it was the top level buildroot directory,
now it is the 'output' directory.
As an side effect, the download directory 'dl' was moved to 'output/dl'.

In commit 416323e93442646e344a4e914776520c5e552fa0 the default download
directory was moved back to the top level of buildroot.

So the various default configurations should reflect this. Fix them.

Signed-off-by: Michael Roth <mroth@nessie.de>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>