package/coreutils: ensure configure isn't considered out-of-date
We were patching m4/rename.m4 to workaround an upstream issue, but this
triggers a auto* rebuild and a configure rerun when we build coreutils
using whatever auto* versions the user has installed.

Doing a manual autoreconf run after patching is unfortunately not an
option as the coreutils configure.ac isn't compatible with the autotools
version we have in BR.

Instead, simply cheat by patching configure as well and setting the
timestamp of m4/rename.m4 sufficiently far back to ensure make doesn't
consider ./configure out of date.

Long term we should convert coreutils to Makefile.autotools.in format,
but this is good enought for 2009.11.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>