xfont_font-util: install in staging as it used to be
Before the autotools conversion, xfont_font-util was installed in
staging, so keep this.

However, the current way of handling xfont_font-util is ugly. The
post-install.sh script removes the binaries compiled for the target
from the staging directory and replaces them by symbolic links to
utils installed on the host (making the incorrect assumption that they
are available).

It looks like xfont_font-util needs to be built for the host (using
the new infrastructure for host autotools package). This is certainly
post-2010.02 though.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>