squashfs: pass CFLAGS/LDFLAGS and add host-zlib dependency
Squashfs depends on host-zlib, so we add this as a dependency. As
host-zlib is installed in $(HOST_DIR), we must pass $(HOST_CFLAGS) and
$(HOST_LDFLAGS).

Unfortunately, if we pass CFLAGS=$(HOST_CFLAGS), we override the
CFLAGS defined in the Makefile, and the build fails. Therefore, we
borrow a patch from OpenEmbedded to fix this problem.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>