netsnmp: disable build with external toolchain
There is a build problem with netsnmp on external toolchain that isn't
easy to fix: libtool strips the --sysroot that we are passing it,
which make the link step fail.

In preparation for the release, just make sure this package isn't
visible to external toolchain users. Of course, on the long run, we
should fix this.

We also disable the build of the Quagga SNMP support because it
selects netsnmp.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>