alsa-lib: make sure to only look in STAGING_DIR for Python includes
Even though we pass a -I option to tell alsa-lib to look for Python
includes in the STAGING_DIR, alsa-lib build process still looks in
/usr/include for some stuff, causing bug #321.

This fix is the one suggested by Ulf Samuelsson in bug #321, and
allows to make alsa-lib with Python support to build properly on a 64
bits machine with python-dev installed on the host. Without this fix,
the build fails with exactly the same problem that Ulf reported.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>