xkeyboard-config: bump to 2.0 to fix bug 2965
Closes #2965

For some odd reason, xkeyboard-config < 1.8 was creating a symbolic
link from /usr/share/X11/xkb/xkbcomp to the xkbcomp binary. But in
cross-compilation mode, this is absurd as the xkbcomp binary to which
the link is pointing is the one in $(HOST_DIR).

This symbolic link thing has been removed completely starting from
xkeyboard-config 1.9. See
http://cgit.freedesktop.org/xkeyboard-config/commit/?id=f413dff57e77e7b01461508f74b4e92d1dc8defd.

Therefore, we simply bump xkeyboard-config to the latest available
version, 2.0.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>