autotools: add escaping when using HOST_MAKE_ENV and TARGET_MAKE_ENV
Without this escaping, the evaluation of HOST_MAKE_ENV and
TARGET_MAKE_ENV is done when the AUTOTARGETS_INNER macro is defined,
not when it is instantiated by the various packages. The result is
that the $(PKG_CONFIG_HOST_BINARY) variable, used in $(HOST_MAKE_ENV)
gets expanded before it is defined.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>