qt: tune .pc files after installation in staging directory
The Qt configuration and installation process is a bit strange, and it
leads the .pc files to contain absolute paths to the staging
directory, while it should only contains paths relative to the staging
directory, because pkg-config already adds the path of the sysroot
automatically to the headers and libraries locations.

This allows the 'pkg-config --cflags' results for Qt libraries to be
correct.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>