bluez-utils: disable on avr32
The bluez-utils package does not build for the avr32 architecture; this
results in autobuilder failures such as the following.

  http://autobuild.buildroot.net/results/d737d1220d7566289eee802fd580a49d8c456c97/build-end.log

The underlying problem is with the <sys/epoll.h> header provided by the
antiquated avr32 toolchain; this header lacks definitions for
epoll_create1() and EPOLL_CLOEXEC.

This patch disables support for the bluez-utils package on the avr32
architecture.

Signed-off-by: Simon Dawson <spdawson@gmail.com>
Acked-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>