libglib2: don't try to use DTrace/GCov/SystemTap for host build
We really only need the host build of libglib2 for a few helper
programs that are used for the target build of libglib2. Therefore,
trying to use bells and whistles like DTrace, GCov and SystemTap is
totally useless.

And it is actually harmful, since it is causing build failures on the
gcc110 PowerPC-based autobuilder that apparently has some
DTrace-feature installed. This commit therefore fixes:

  http://autobuild.buildroot.org/results/1c62d1ce10937bd21f5afcb73782b939d10c2038/build-end.log

Which has been polluting our autobuilder logs since several weeks now.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>