linux-fusion: don't override CFLAGS/LDFLAGS
Putting $(TARGET_CONFIGURE_OPTS) after $(MAKE) overrides the Makefile's
CFLAGS and LDFLAGS. Some of these flags are required, however. So
instead pass these things in the environment, which allows the Makefile
to append to the flags.

This removes the need for the patch, because now the correct -I options
are added.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Tested-by: Samuel Martin <s.martin49@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>