xlib_libX11: do not depend on bigreqsproto nor xcmiscproto
The BR2_PACKAGE_XPROTO_BIGREQSPROTO option was selected by
xlib_libX11, but xlib_libX11 does not have this protocol package in
its DEPENDENCIES. In some builds, it leads to xproto_bigreqsproto to
be built as the last package, with no other package depending on it,
which doesn't make sense.

xproto_xcmiscproto was selected and part of DEPENDENCIES, but an
inspection of libX11 configure.ac and libX11 source code shows that it
does not depend on xcmiscproto at all.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>