New package : GNU SASL Library - libgsasl
GNU SASL is an implementation of the Simple Authentication and Security
Layer framework and a few common SASL mechanisms.  SASL is used by network
servers (e.g., IMAP, SMTP) to request authentication from clients, and in
clients to authenticate against servers.

libgsasl will be used in future versions of libesmtp.

[Peter: Misc minor fixes]
Signed-off-by: Sagaert Johan <sagaert.johan@skynet.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>