autoconf: don't try to build Emacs files
When building autoconf for the target, we already passed EMACS="no" to
prevent autoconf from building Emacs mode files. But we weren't doing
that when building autoconf for the host. This causes problems when
'emacs' is not really emacs, but a sort of clone like Jove. So we also
pass EMACS="no" when building host-autoconf to avoid autoconf
./configure script from detecting emacs and then use it to build .elc
files from .el source code.

Reported-by: Spielmann Werner <Werner.Spielmann@swarovski.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>