integrator926: Update the defconfig
the integrator926 defconfig has not been updated for quite some time,
and is still asking for very old linux and u-boot releases. Now that
these versions aren't present in Buildroot anymore, they default to a
newer release, where, in the u-boot case, the defconfig name has
changed, leading to a build failure.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>