json-c: bump to version 0.11
Bumping this package is needed, because json-c 0.10 has a major
defect: it installs a library named libjson.so, which conflicts with
the library installed by the libjson package.

This has been changed in the upstream json-c 0.11 version, which now
installs libjson-c and json-c.pc.

It allows to solve the Pulseaudio link problem, which wants to link
against json-c but happens to link against libjson when libjson has
been compiled and installed after json-c.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>