xlib_libX11: ensure target and host variant have the same patch hooks
In preparation for the out-of-tree support, it is important that all
packages have the same patch hooks for the host variant and the target
variant (because the source tree will be shared between host and
target variants when doing out-of-tree build).

In the case of the xlib_libX11 package, the post-patch hook removes
some flags from the build of the makekeys utility, so that when
xlib_libX11 is built for the target, the build of makekeys for the
host works properly. This modification can be applied to the host
variant of xlib_libX11 without any problem.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>