package/linux-firmware: bump version for new firmware
Bump version to get a few new firmware files for some WiFi adapters.

In the rtl81xx familly, it's hard to differentiate each members: I have
a USB stick that identifies itself as TRL8188CUS, but is handled by the
rtl8192ce driver. So I moved all of rtl81xx firmwares under the same
option to avoid confusion.

It will be the responsibility of the user to eventually get rid of
unwanted firmware files in a post-build script, if needed.

Ditto for rtl87xx.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>