package/libedit: switch to alternate URL for proper packaging
Upstream libedit does not have a proper homepage, and does not provide
any sort of packaging whatsoever. So far, we used Debian's wayback
machine to get a tarball of a rather oldish libedit.

As Thomas pointed out, someone has been maintaining a proper autotools
based, up-to-date package for a while.

While Debian is stuck to 2.11 (5 years old now), we bump to 3.1 as a bonus.

Remove our libedit.pc, it's now bundled with the package.

Reported-by: Thomas De Schampheleire <patrickdepinguin+buildroot@gmail.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas De Schampheleire <patrickdepinguin+buildroot@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>