boost: fix 64-bit build due to ambiguous format call
This commit adds a patch imported from the upstream bug tracker, fixing a
problem with boost builds failing for 64-bit, like in autobuild [1].
Refer to [2] for the actual autobuild error.

Upstream did not yet accept this patch, because it is not portable enough
(not OK for Windows builds) [3], but this should not withhold us from fixing
the error in buildroot.

[1] http://autobuild.buildroot.net/results/0f6c2f99b05f57adb557c2ddd686b5072f5e9ecf/
[2] http://lists.busybox.net/pipermail/buildroot/2013-August/076222.html
[3] https://svn.boost.org/trac/boost/ticket/6851

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>