package/weston: add RPi compositor
Procedure highly inspired by:
    http://wayland.freedesktop.org/raspberrypi.html

The resulting weston works almost flawlessly, but requires a bit
of love:
  - /boot/config.txt must include this line: dispmanx_offline=1
  - at least 128MiB of RAM must be allocated to the GPU
  - after 24-or-so terminal-clients are connected, the screen
    turns black. Exiting a client restores the screen

It seems increasing/decreasing the amount of memory allocated to
the GPU makes the clients limit to wobble above/below 24 clients
at a time. YMMV, as they say...

Without dispmanx_offline=1, the limit is much below 24, at around 13.
But changing the amount of memory allocated to the GPU does not change
this limit in this case. YMMV, again.

Anyway, there are not many different clients available, besides the
terminal client, since all other clients are EGL-based, and there
is (yet) no EGL support (for weston!) on the RPi. So the tests were
made only with the terminal client.

The system is rather smooth, but spwaning too many clients in a
rapid-fire is sure to exhibit some lag. Resizing windows is a bit
jerky, but moving them along is fine.

Note: the config option has a depends on THREADS due to rpi-userland,
even though weston itself already inherits the same dependency from
wayland. But better be clean and safe.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>