docker: change into generic-package
docker is using autotools-package while it is in fact just a plain make
package.

Remove docker-1.5-skip-configure.patch which was just a hack to avoid
calling the non-existent configure script.

Remove docker-1.5-pkgconfig.patch which is not needed when PATH is set
appropriately.

Add a XLIBPATH option to the command line to avoid linking with
/usr/X11R6/lib.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>