wmctrl: new package
Provides command line access to almost all the features defined in
the EWMH specification. It can be used, for example, to get
information about the window manager, to get a detailed list of
desktops and managed windows, to switch and resize desktops, to
make windows full-screen, always-above or sticky, and to activate,
close, move, resize, maximize and minimize them.

[Thomas:
 - fixed license, which is GPLv2+, not GPLv2
 - fixed wrapping of the Config.in help text
 - removed Config.in comment related to MMU dependency.]

Signed-off-by: Jérôme Pouiller <jezz@sysmic.org>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>