mmc-utils: fix linux headers issue before 3-4
With toolchains that include linux kernel headers before the 3.4 kernel,
an issue exists with linux/mmc/ioctl.h in that it doesn't include
linux/types.h.

Merged the patch with an existing patch for mmc-utils as it modifies
the same file and fixes similar issues. Also updated the patch to
conform to the buildroot patch naming standard.

Fixes:
  http://autobuild.buildroot.net/results/828/828c7a686736710c4fcb152230d0fa6b5f4d6ad3

Signed-off-by: Ryan Barnett <rjbarnet@rockwellcollins.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>