vlc: fix some configure options
Fix a large number of incorrect configure options for vlc.

Configure warnings reported by Arnout Vandecappelle, in his analysis of vlc
autobuild failures.

Notice that this change of configure options is caused by our AUTORECONF
picking up a different PKG_WITH_MODULES definition.  If vlc is bumped in the
future and AUTORECONF is turned off, then this needs to be reverted.

Signed-off-by: Simon Dawson <spdawson@gmail.com>
Reviewed-by: Samuel Martin <s.martin49@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>