libnftnl: exclude for CS PPC toolchains
The CS PPC 2011.3 and 2010.9 toolchains have a bad linux/socket.h header
that doesn't define sa_family_t which is used by linux/netlink.h (and
libnftnl).
Patching external toolchain headers wouldn't be too nice and they seem
to be the only CS toolchains affected.
It's definitely NOT a kernel header version issue, the toolchain is
based on 2.6.38 headers and vanilla 2.6.38 has the proper definitions in
place. Fixes:
http://autobuild.buildroot.net/results/1e2/1e2cac84f0faa9b8915a628260ef41765d86e174/

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>