package/dvb-apps: bump revision
In this new revision, there is no longer transponders data, so dvb-apps
now only installs the dvb-apps utilities.

Packages that needs the transponders data have now all been switched to
use dtv-scan-tables instead, so we don't need a legacy option for the
transponders data.

However, we add a legacy option for the utilities option, just to inform
the user of the new behaviour.

Sadly, a user that had dvb-apps selected just to get the transponders
data will now get the full dvb-apps package installed. There's nothing
we can really do about this.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>