Remove package xdriver_xf86-video-newport
http://patchwork.freedesktop.org/patch/14670/ (State Accepted)
"This driver has not received server updates such as mibstore removal
and conditional compiling of xaa architecture. This indicate an implicite
decisison to remove the driver from active support and future releases.

This driver is used for SGI workstations Indy and Indigo2. Manufacturing
stopped in 1997 and SGI support ended in 2012."

Signed-off-by: Bernd Kuhls <berndkuhls@hotmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>