package/duma: do not let Makefile guess OS
We need to tell the duma Makefile what OS we're gonna run on, or it
uses default values.

It turns out the default values are the same as for OS=linux (and so
this change should be a no-op), but it's just better to force it, and
it avoids a warning during the build.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>