package/perl-net-ssleay: fix buildsystem for cross-compilation
The buildsystem for perl-net-ssleay is totally brain-damaged: it tries
to execute the openssl program. It uses that to check the openssl
version, detect the prefix to openssl, and set the include and library
search paths.

Of course, it does not handle the destdir idiom, so it ends up finding
the host system's openssl (not even ours!), so it sets search paths to
/usr/inlcude and /usr/lib and /lib.

Well, guess what? It does not work.

Fix that by:
  - passing the prefix to openssl in the environment, so it does not go
    hunting for the host-system openssl;
  - not checking the version string, since we can't run, and Buildroot
    has the correct version anyway;
  - not setting any -I or -L paths since our compiler/linker already
    know where to look for for includes and libs.

Fixes:
    http://autobuild.buildroot.net/results/135/135867ef85535863e3647cc5fb82accb6f77612c/

Signed-off-by: Francois Perrad <francois.perrad@gadz.org>
[me: enhance commit log, add patch to avoid running openssl and setting
 wrong include and library search paths]
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>