btrfs-progs: bump to version 3.16, fix static linking
The original aim of this patch was to fix static linking of
btrfs-progs. In order to implement this in the form of patches that
can potentially be upstreamed, this patch first bumps the btrfs-progs
package to the latest upstream stable version. It then:

 - Reworks the existing documentation disabling patch in a form that
   can potentially be upstreamed.

 - Adds a patch to support static building of the btrfs-progs
   binaries.

Both patches have been submitted upstream, see
http://article.gmane.org/gmane.comp.file-systems.btrfs/38145.

Fixes:

  http://autobuild.buildroot.org/results/ddc/ddcc70143ce2c2882894184e4ce195a11407e027/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>