ffmpeg: add --pkg-config to configure options
The ffmpeg does not understand the PKG_CONFIG env variable,
so it throws a line like this at the end of its configuration:

WARNING: .../output/host/usr/bin/arm-none-linux-gnueabi-pkg-config not found, library detection may fail.

Apparently, some of the indevs and outdevs need a pkg-config.
This patch adds the --pkg-config to let ffmpeg know where it
is.

[Thomas: remove unrelated formatting change.]

Signed-off-by: Danomi Manchego <danomimanchego123@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>