package/linux-firmware: split bcm43xx / bcm43xxx
linux-firmware has gained a few firmware files for Broadcom WiFi
chipsets. Installing all of them takes a lot of place, when usually only
one is really needed.

Split the Broadcom BRCM drivers in two categories: bcm43xx and bcm43xxx,
when they were previously a single config option, so as to reclaim a bit
of space.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>