bandwidthd: Version bump to handle indirect deps properly
Thanks to Romain Naour, Baruch Siach and Thomas Petazzoni, I've
solved the build issues in certain cases for bandwidthd.

The issue was that libpcap and libpng would depend on other libraries
and upon testing for pcap/libpng the tests would fail due to not
being able to find the other libraries.

Bandwidthd's configure script now uses pkg-config for libpng
and pcap-config for libpcap. pcap-config's path is specified
in configure's arguments.

Some other tweaks have been made including the layout and flow
of configure.ac is a bit more logical, and the handling of postgres
and sqlite has been tweaked. Finally, the test for X.Org has been
removed.

[Thomas: re-add Nathaniel SoB line, which was misplaced in the commit
log, and therefore dropped when applying the patch.]

Fixes: http://autobuild.buildroot.org/results/eb4/eb48acec37a7fd1714b267b105816bf0053d40e9/build-end.log
Cc: Baruch Siach <baruch@tkos.co.il>
Cc: Romain Naour <romain.naour@openwide.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Nathaniel Roach <nroach44@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>