package: prepare to support modules that work with Qt or Qt5
Currently, all the Qt external modules (qtuio, qjson, etc.) can only
be compiled against Qt4. However, some of them (qjson, qextserialport
at least) can compile with either Qt4 or Qt5.

Therefore, this commit adjusts the Config.in logic to make it possible
for certain modules to be selected with either Qt4 or Qt5. We use the
same approach as the one of for Python 2 vs. Python 3 modules:

 - in package/Config.in, the package Config.in files are included
   either if BR2_PACKAGE_QT or BR2_PACKAGE_QT5 are enabled.

 - each individual package is then responsible for having a dependency
   on either BR2_PACKAGE_QT, or BR2_PACKAGE_QT5 or both depending on
   what they support. In this commit, we add a dependency on
   BR2_PACKAGE_QT to all packages, since for now, they only support
   Qt4.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>