host-openssl: use correct prefix so libssl / libcrypto can be found
Fixes (root cause of):
http://autobuild.buildroot.net/results/d01/d0190a6e5c6b5aa78cfefc29e4e8a4e4b135450b/
http://autobuild.buildroot.net/results/9bb/9bb8defedb8024de83eb3a609ab584efb5d0d2b0/
http://autobuild.buildroot.net/results/e4a/e4a1b298e7e158d3614b89e61d40ef9a00e7ff67/
http://autobuild.buildroot.net/results/5e2/5e24a90058111f7d5618047410f809b45bc165f6/

And many others.

The host version of openssl should be configured to use the host directory
as its prefix and the INSTALL_PREFIX dance shouldn't be used, as otherwise
host-openssl will look in /usr for its libraries.

While we are at it, ensure our host CFLAGS are used similar to how we do for
the target build.  Also enable zlib support as we already list it as a
dependency.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>