package/shairport-sync: fix init script
shairport-sync uses libdaemon for running as daemon and creating the pidfile
which conflicts with start-stop-daemons `-b -m` options. For that reason
shairport-sync does not start using /etc/init.d/S99shairport-sync start.

The issue is fixed by removing the background and pidfile task from
start-stop-daemon and add a daemon option `-d` to shairport-sync.

Signed-off-by: Jörg Krause <jkrause@posteo.de>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>