package/network-manager: NM needs isc-dhcpc or dhcpcd
NM only currently supports isc-dhcpc and dhcpcd as clients because NM
communicates with the client through DBus, and support for clients needs to
be written in.  (as per files in src/dhcp-manager)

Buildroot's default, udhcpc is not supported.

Add the dependencies to Config.in, favoring dhcpcd as
default due to size.

[Peter: drop references to dhclient, isc-dhcpc is dhclient]
Signed-off-by: Nathaniel Roach <nroach44@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>