rng-tools: bump to version 5
Add semi-automatic libgcrypt dependency for i386/x86_64 to use RDRAND.
This is exclusively used by rngd for which we don't yet ship an
initscript/service file. Even so the user can choose to disable it at
invocation time via the --no-drng option (which must be accounted for).

Also add hash file.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>