package/mutt: specify the mailpath
mutt's ./configure looks for different paths for where to look for
incoming mails.

This is absolutely worng in the case of cross-compilation, because the
path it may find on the host may not exist on the target.

Not only that, but some host may not even have any of the paths
./configure looks for.

Fix that by specifying the mailpath, and set it to the value documented
in the FHS [0]

Since Buildroot only guarantees /tmp to be writable, make /var/mail a
symlink to /tmp .

[0] http://refspecs.linuxfoundation.org/FHS_2.3/fhs-2.3.html#VARMAILUSERMAILBOXFILES

Reported-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>