xlib_libXau: add host-pkgconf to its dependencies
xlib_libXau needs pkg-config during the configure phase. Otherwise it
will fail with an error like this one:

checking for XAU... configure: error: in
`/br/output/build/xlib_libXau-1.0.8':
configure: error: The pkg-config script could not be found or is too
old.  Make sure it
is in your PATH or set the PKG_CONFIG environment variable to the full
path to pkg-config.

Signed-off-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>