ola: bump to 0.9.3
This update requires a few changes. First, good news, the patches that you
guys have submitted to OLA have been merged, so we can drop those:

 ola-0001-fix-build-warning.patch -> ea375582b0bfee93d66608ffc807078ffc48e961
 ola-0002-move-python-sub-check-to-configure.ac.patch -> 673a7602a6bf7a6aa8a8461ebd9362d59f6e21df
 ola-0003-fix-check-for-python-module-for-cross-compilation.patch -> b51b48be81ec38bc7d1229be0c7d3189c5ddbafa

Less good news: OLA now builds a small protoc wrapper (ola_protoc) that is
built and executed at compile-time on the host. If we don't change anything,
ola_protoc is built with the target toolchain and therefore can't run on the
host. Explanation for ola_protoc is here

To solve this, I created a package host-ola, which builds and provides an
ola_protoc for the host. It tries to disable as much as possible of things
that we won't need at configure time. Only ola_protoc is built and installed
so it's not that long.

The change has been built-tested only.

[1] https://github.com/OpenLightingProject/ola/blob/master/protoc/ola-protoc.cpp#L20

[Thomas:
 - add HOST_OLA_DEPENDENCIES variable, so that host-ola doesn't get
   all the dependencies of the target ola package: certain target ola
   dependencies do not have corresponding host packages (and it
   doesn't make sense to have them as dependencies just to build
   ola_protoc)
 - improve the commit log to mention which upstream commits correspond
   to our patches.]

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>