memtest86+: new package
Memtest86+ is a bootable standalone memory test program.

Please note that this is the forked memtest86+ program and not
the original memtest86 which has different licensing. Buildroot
does not support packages with a '+' sign in their name.

Memtest86+ is a utility designed to test whether your memory
is in working order. It repeatedly writes an enormous amount
of different patterns to all memory locations and reads them
back again and verifies whether the result of the read is the
same as what was written to memory.

Memtest86+ will only work on 32-bit or 64-bit x86 targets.
It boots as an i486 program and autodetects hardware.

[Peter: tweak help text as suggested by Thomas]
Signed-off-by Stephen M. Kenton <skenton@ou.edu>
Reviewed-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>