rpm: use pkg-config to detect zlib and popt
Fixes:
http://autobuild.buildroot.net/results/98b/98b70ea09ecef214f13a622dcb926e01a6d6f7d9/
http://autobuild.buildroot.net/results/d13/d1343c3be9710d5be3cf6e4369dd5c75102e7b92/
http://autobuild.buildroot.net/results/457/457d3d9231f1422b19ffbac6d4a6ab02c6d6c22f/
http://autobuild.buildroot.net/results/ed8/ed8efe3985e6b423e4c18f9c200aa042ccb9ed23/

And many more.

Using --with-zlib=$(STAGING_DIR) causes the configure script to basically do
a find $(STAGING_DIR) -name zlib.h to figure out the correct include path
for zlib.  That is normally ok, but if botan is enabled and built before
rpm, then it installs a zlib.h header in
$(STAGING_DIR)/usr/include/botan-1.10/botan which confuses this logic and
causes configure to fail.

Fix it by using --with-zlib=external instead which causes the configure
script to use pkg-config to get the CFLAGS/LDFLAGS. Also do it for popt so
we don't end up with a similar issue in the future. We unfortunately cannot
do it for beecrypt as it doesn't install a .pc file.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>