package/nvidia-driver: add NVidia's OpenGL binary blob
This patch only adds the userland part. Unless other such other
packages (which we named like: rpi-userland), we do not replicate this
naming scheme with this package, as a future patch will also enable
building the kernel part of the driver. So, it is better to just name
that package with -driver, rather than with -userland and renaming it
afterwards.

[Thomas:
 - Rewrap Config.in help text.
 - Add a comment to explain why mesa3d-headers, xlib_libX11 and
   xlib_libXext are part of the dependencies.
 - Fix typo in comment about library installation: s/The/Then/
 - Use 'addsuffix' instead of 'patsubst' to calculate the final
   filename of libraries to install.
 - Use more temporary variables to make the library installation loop
   clearer: 'libpath' is the relative path of the library in
   nvidia-driver sources, 'libname' the base name of the library,
   'libsoname' the soname of the library, and 'baseso' the base .so
   symlink name.]

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>