package/thrift: force boost libdir to staging/usr/lib/
Libtool append the sysroot path (lt_sysroot) with the libdir value found in
the libstdc++.la file only with bfin toolchain.

Path/to/sysroot/opt/uClinux/bfin-linux-uclibc/lib/gcc/bfin-linux-uclibc/4.3.5/libstdc++.so

The issue disappears when --with-boost=$(STAGING_DIR) is removed.
So instead of using --with-boost=$(STAGING_DIR), use
--with-boost-libdir=$(STAGING_DIR)/usr/lib to give the path to boost's
libraries.

Fixes:
http://autobuild.buildroot.net/results/ae2/ae274693b953dd569c3ff63c532e2fc9553b5002/

And many more

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>