sane-backends: bump to version 1.0.24
Bump sane-backends to a "modern" 1.0.24 version.
Tweak the patches to avoid autoreconf (it doesn't seem to be happy for
me in gentoo).
Disable cups support (checks are wrong and it's no longer required).
Disable systemd support (checks are wrong and we never accounted for
this).

Should fix wine failure (i was unable to build sane-backends before):
http://autobuild.buildroot.net/results/d01/d01a3c7786df8cfd962a27f39283851db45c6165/

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>