ipmiutil: force md2 off
The configure test runs against LIB_DIR (which is --libdir in configure
parlay) hence it's against the distro openssl version.
On newish openssl versions md2 is disabled by default, and buildroot
doesn't try to enable it (it's obsolete), but if the distro
libcrypto/openssl has it enabled then ipmiutil tries to use it.

Force it off to avoid build breakage. Fixes:
http://autobuild.buildroot.net/results/cd2/cd2e617f8e2b00581ab5936029f85e62ed3259ba/

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>