libnftnl: needs kernel headers >= 3.1
Contrary to what commit f5866484f (libnftnl: exclude for CS PPC toolchains)
stated, a missing sa_family_t definition IS specific to the kernel headers
version, and not something particular to the CS toolchains, so instead
require 3.1+ headers.

The specific kernel change fixing the headers is:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=6602a4baf4d1a73cc

The only reverse dependency of libnftnl (nftables) already needs 3.4+, so
this dependency doesn't need to be propagated.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>