lua-periphery: bump version to 1.0.5, remove dependency on c-periphery
The new version of the lua-periphery LuaRocks package handles fetching the
c-periphery dependency, so that it no longer needs to be cloned during the
lua-periphery build process. Consequently, this removes the need for selecting
c-periphery as a package dependency in buildroot and the need for the patches
that modify lua-periphery to be built against an external c-periphery.

[Thomas:
  - remove dependency from .mk file, as noticed by Baruch
  - add patch to fix uClibc build]

Signed-off-by: Vanya Sergeev <vsergeev@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>