package/lttng-modules: use kernel-module helper
lttng-modules' Makefile already uses the Linux buildsystem to build its
modules, so just switch to using the new kernel-module infra.

To be noted: lttng-modules' Makefile uses $(shell pwd) that expects the
PWD to be the Linux' top build directory, which is not the case for us.
However, it uses to detect a conflict between a stale static version.h
and a generated version.h (due to the switch to uapi). This is of no
consequence to us, because we can't have a Linux build directory with
such stale files.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>