ltp-testsuite: support building with toolchains without native RPC
ltp-testsuite needs RPC, but this could also be provided by libtirpc.

The dependency of libtirpc on
!BR2_TOOLCHAIN_EXTERNAL_BLACKFIN_UCLINUX_2012R2 does not need to be
propagated, because it is always satisfied: ltp-testsuite depends on
BR2_USE_MMU which is always unset for Blackfin targets.

Since musl toolchains never have RPC support, this change would now allow
building of ltp-testsuite on musl toolchains. Unfortunately, ltp-testsuite
does not build yet with musl, so a specific check on musl is added.
This is deemed more conceptually correct than checking on glibc||uclibc.

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Cc: "Yann E. Morin" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>