lrzsz: mark as not available for static builds
lrzsz could technically build in a static-only environment, but it
unconditionally redefines the error() and error_at_line() functions,
which clashes with the definition from uClibc. Since the autotools
packaging dates back from 1998, lots of work is needed to fix it, and
do some proper tests on the availability of such functions.

So, let's mark it as not available for static builds, until someone
interested enough fixes it.

Fixes:

  http://autobuild.buildroot.net/results/8106ab057f538d1408d736736fdbe50b99ae709d/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>