zyre: bump version to fix static linking problem
The last tagged release of zyre, v1.0.0, was made in May 2014. Since
then, they have switched to pkg-config to detect zmq and czmq, which
fixes static linking problems. However, they made a number of changes
to configure.ac, which make it difficult to backport just the relevant
changes.

Since we already had another backported fix, let's simply bump the
version of zyre to the latest available commit, which builds fine with
no change for shared and static scenarios, thanks to the use of
pkg-config.

An issue was opened upstream to ask them to tag a new release:
https://github.com/zeromq/zyre/issues/324.

Fixes:

  http://autobuild.buildroot.net/results/0ab/0ab4c6a4bb4942d51e7712073d4731d81ecb5251/

Thanks to Vincente Olivert Riera for the initial investigation.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>