xdriver_xf86-video-imx-viv: bump to version 5.0.11 p4.5
Modify the version to match the version of imx-gpu-viv. Also add the
package hash file.

This patch is based on the Yocto equivalent:
https://github.com/Freescale/meta-freescale/commit/45582680

The following xorg.conf was used in order to force the use of the
vivante module:
https://github.com/Freescale/meta-fsl-arm/blob/master/recipes-graphics/xorg-xserver/xserver-xf86-config/mx6/xorg.conf

Note that the X server must be started with the "noreset" option to
avoid crashes when closing X apps.
 # X -noreset

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>