configs: update cubieboard_defconfig to a 3.4 kernel
The current commit used for the kernel Git repository for Cubieboard,
274a66a7bfcbaabb88d63e4eba161965383cc416, actually points to 3.0
kernel. This is incorrect since the kernel headers version specified
by the defconfig is 3.4. And anyway, the 3.4 kernel is kind of the
official vendor kernel for Allwinner platforms.

mfld.fr@gmail.com reported in bug #7931 that commit
9a1cd034181af628d4145202289e1993c1687db6 was working for him. It is
the latest commit in the sunxi-3.4 branch. So we switch to using this
commit, which was build tested successfully.

While we're at it, remove the definition of BR2_LINUX_KERNEL_VERSION
from the defconfig: it doesn't make sense to have it defined in a
defconfig, as it's a blind option that gets defined from the value of
BR2_LINUX_KERNEL_CUSTOM_REPO_VERSION when a Git repository is used to
source the kernel.

Fixes #7931.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>