liquid-dsp: fix powerpc/altivec build failures
When a powerpc target is detected the build system tries to force
altivec support but that's wrong:

Not every powerpc processor has altivec instruction support.

It's using the removed/deprecated -faltivec compiler directive when it
should be using -maltivec instead.

Even if the above conditions are corrected the altivec codepath fails
to build properly.

So just disable altivec in general using the standard
--enable-simdoverride configure option. Fixes:

http://autobuild.buildroot.net/results/e1b/e1b6cf76a7a4ae754a928d211e779e9381198f48/

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>