libserial: add patch to fix build failure on sparc/sparc64
Like c-periphery and lua-periphery, libserial fails to build because
it tries to use some baud rate definitions that are not available on
sparc and sparc64 (the highest baud rates).

This commit fixes that by introducing a patch to libserial to make the
use of these high baud rates conditional on their availability.

Fixes:

  http://autobuild.buildroot.org/results/f9b/f9bbb8a6636cd3e3203b059f627aac7b1d511eb2/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>