rpi-userland: fix musl compile
When building with musl, the build fails with a conflicting type
qualifier error.

This is because musl defines stdout and stderr as const pointers
and RaspiVidYUV.c adds an extern declaration as normal pointers.

Since this declaration already comes in on the header (stdio.h),
there is no need to add an extern declaration, so removes it.

Fixes:

http://autobuild.buildroot.net/results/27bd6f32cecdb4e7c95247c1feaf2732c1d8e3fa

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>