trousers: disable pie option on ARC
ARC gcc understands "-pie" option and attempts to generate PIE
binaries as of today PIE is not really supported for user-space
applications.
So we provide option which checks if compiler understands relro
and pie options and disable PIE detection if building for ARC.
Also AUTORECONF option should be added because of modified
configure.in and Makefile.am files.

Signed-off-by: Lada Trimasova <ltrimas@synopsys.com>
Cc: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Lada Trimasova <ltrimas@synopsys.com<mailto:ltrimas@synopsys.com>>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>