linknx: argp-standalone also needed with musl
The linknx package needs the argp functionality, which is provided
built-in by glibc, but not by uClibc and musl. In uClibc
configurations, it correctly selects the argp-standalone package to
compensate this, but forgets to do so in musl-based
configurations. This commit fixes that by selecting the
argp-standalone package.

Found by checking all packages having an argp-standalone
dependency. Note that this commit is not sufficient to make linknx
build properly with musl.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>