vpnc: switch to github as the new site
The vpnc upstream project has not made any tarball release since many
years, but they have continued to update their Subversion
repository.

A Github repository synced with the Subversion repository now exists
at https://github.com/ndpgroup/vpnc, which is more practical to
generate patches and submit pull requests. For this reason, we move to
using this Github repository as the new source for the vpnc
package. The specific commit that we use corresponds to version 0.5.3
exactly.

All patches are re-formatted using Git.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>