zeromq: unbreak pgm support
In the bump to the 4.1.x series the bundled pgm library was dropped and
the --with-system-pgm option was made pointless since using unbundled
became the only option, so it was renamed to --with-pgm, which
previously meant "use bundled" (but not any longer).
However this wasn't accounted for and pgm support has been broken since
then.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>