libepoxy: move outside x11r7 scope
Move it outside x11r7 scope since it can be built without it with a few
patches (sent upstream via github pull request).

Motivation is simple, it's a dependency for newer gtk3 versions which
can work with a wayland or broadway backend, and having a full x11 stack
is pointless for that scenario.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
[Thomas: split the first patch into four separate patches, since
that's how they were submitted upstream.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>