xdriver_xf86-video-imx-viv: bump to version 5.0.11 p7.4
Changelog:
- License changed to MIT.
- Fix GCC5.2 build.

Also add a patch pulled from Yocto for the following reason:
The 'swaprep.h' is an internal Xorg header and is not being installed
anymore. The swap features are provided by the 'misc.h' header now.

This patch is based on the Yocto equivalent:
https://github.com/Freescale/meta-fsl-arm/commit/19935f57
https://github.com/Freescale/meta-fsl-arm/commit/277cef92

The following xorg.conf was used in order to force the use of the
vivante module:
https://github.com/Freescale/meta-fsl-arm/blob/master/recipes-graphics/xorg-xserver/xserver-xf86-config/mx6/xorg.conf

Note that the X server must be started with the "noreset" option to
avoid crashes when closing X apps.
 # X -noreset

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>