package/qwt: prepare package to support other versions of Qt
Rework the selection of Qt options in the qwt package to make it
easier to support other versions of Qt than just Qt4. Instead of
depending on Qt options, we select the ones we need, and simply have a
global "depends on BR2_PACKAGE_QT". This also allows to remove a
comment that becomes useless.

Signed-off-by: David Picard <davepiq@yahoo.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>