netsnmp: ensure lm-sensors is available if enabled
The ucd-snmp/lmsensorsMib module uses lm-sensors, so ensure netsnmp is built
after lm-sensors if enabled.

As the list of mib modules to built is configured using a string option in
Kconfig, we cannot easily select lm-sensors from the netsnmp package - But
we can at least ensure it gets built before hand if enabled.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>