yajl: remove obsolete patch
Commit 302563539dacb284576a443401cdfd061eb2e1e8 fixes link with libm for all
targets linking against libyajl{,_s}. This makes patch "0001-cmake-uClibc-Fix
-missing-libm-for-tests" obsolete so it is safe to remove it.

Instead of renumbering all remaining patches the patch defined in YAJL_PATCH
is fetched and put intto the packages directory. Having all the patches in one
places makes the number and order of the patches applied to the package more
clear.

Align patch "0003-Link-with-shared-libyajl-in-a-shared-build" with commit
302563539dacb284576a443401cdfd061eb2e1e8 and remove linking with libm from
test/api/CMakeLists.txt.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>