python-crossbar: fix byte-compilation of .pyc files
For its continuous integration tests, python-crossbar includes one .py
file that intentionally has a syntax error. Until now, this wasn't
causing any problem, but with the introduction of a Python
byte-compilation process that errors out when there is a failure, it
is going to become a problem.

We simply add a patch that removes the syntax error, as we don't care
about it in the context of Buildroot. The problem has been reported
upstream to see if a better fix is possible.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Samuel Martin <s.martin49@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>