python-pygame: bump version to fix Python 3 byte-compilation issue
The current version of python-pygame in Buildroot does not
byte-compile properly with Python 3. Commit 22efec0b44da from upstream
fixes the problem, so we take this opportunity to bump the pygame
version.

Fixing byte-compilation is necessary as we will soon enable a
mechanism that aborts the build in case of error during the
byte-compilation process (while they are today considered as
warnings).

In order for the bump to work, we explicitly disable freetype support,
as the package doesn't handle the freetype dependency for the moment.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Samuel Martin <s.martin49@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>