package/ipsec-tools: disable for musl toolchains
__P() is used for compatibility with old K&R C compilers. With
ANSI C this macro has no effect.

Unlike for util-linux and ipkg packages where it was easy to remove
each __P() macro, ipsec-tools use it all over the tree and require a
"big" patch to enable musl support.

Since upstream seems not verry active (last release 2014-02-27)

So, disable ipsec-tools with musl based toolchains.

This fixes a compilation error with musl libc because of undeclared
__P.

Fixes:
http://autobuild.buildroot.net/results/42242e3f4485b9e77a916e6fe480c83f70e024e4

While at it, reorder "depends on" and "select" lines in Config.in

Signed-off-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>