xserver_xorg-server: apply modesettings patch for 1.18.3 only
We currently have two choices for the server's version (1.18.3, and 1.14.7)
but the following patch will only apply for versions >= 1.17.2.

If 1.14.7 is selected the build will fail with the following error:

Applying 0001-modesettings-needs-dri2.patch using patch:
patching file configure.ac
Hunk #1 FAILED at 2036

So move this patch to an appropriate version folder in the package's
directory. It should be good enough until this package provides more
versions >= 1.17.2.

Signed-off-by: Stanislas Bach <sbach@0g.re>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>