package/skeleton: respect variables namespace
In makefiles, variables have global scope. We traditionally ensure that
we get no name clashing by prefixing all variables with the package
name.

Currently, this is not the case in the skeleton package, for historical
reasons (code snippets copied over from the common Makefiles). We
currently have a mix of naming for the variables:
  - some are indeed prefixed with SKELETON_
  - some are prefixed with SYSTEM_
  - some are prefixed with both
  - some are not prefixed

Clean up these discrepancies, and prefix all variables with just
SKELETON_ and drop the SYSTEM_ prefix.

Also include SET_ in all variables that do set something (getty) for
consistency across all the variables.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>