package/perl: use dummy hostname
The hostname does not look like it serves any useful purpose, except
maybe to set set perladmin email.

Which is undoubtfully useless on the target.

A followup commit will make the hostname depend on the default skeleton,
so it won't always be available. We can not rely on it to be set.

Besides, even today it is not guaranteed to be set; a user may well
leave it empty.

Use a dummy hostname.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>