package/mplayer: needs BR2_TOOLCHAIN_HAS_SYNC_4
Even though mplayer bundles the ffmpeg code base (and ffmpeg is not
affected by this issue), mplayer uses its own build system and doesn't
test properly for the availability (or not) of atomic operations. In
order to keep things simple, we simply make mplayer depend on the
availability of atomic operations, since it assumes they are available.

The reverse dependency is propagated to libplayer and tovid.

Fixes:

  http://autobuild.buildroot.net/results/6306757da3be9bb7e3cdcbfc8abb4e64a3fb7913/

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>