package/efl: disable libevas jp2k loader
When the support for JP2K is requested, the configure script is looking
for libopenjpeg1.pc which is not provided by the openjpeg package since
it's too recent (currently v2.1).

Mark BR2_PACKAGE_EFL_JP2K as broken until a new version of the efl
stack support the new openjpeg library.

Fixes:
http://autobuild.buildroot.net/results/ff1/ff1dbbc6cffdb5b2550b3613295151ff9fa964d6

Signed-off-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>