package/kodi: Add suboption to control MySQL support
In a Buildroot configuration with mysql enabled it might be useful to
be able to build Kodi without mysql support because Kodi may not need
it at runtime.

Kodi can store its internal databases not only locally using sqlite
but also on a mysql server, this allows several local Kodi machines to
share the same databases. When using only one Kodi instance, mysql
support is not needed, and not building mysql support helps reducing
build time.

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>