system: zoneinfo usable by musl, too
Currently, we limit installing zoneinfo to non-musl toolchain, by lack
of knowledge on how it would work on musl.

Turns out that musl uses the same zoneinfo format as glibc does.

Make it possible to install the TZ info whatever the C library; for
musl, use tzdata as for glibc.

Thanks Rich! ;-)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Rich Felker <dalias@aerifal.cx>
Cc: Marc Khouri <marc@khouri.ca>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>