gcc: disable C++ support for Blackfin
As discussed with Waldemar, the C++ support for Blackfin is currently
broken, and we don't have a fix in sight for the 2016.08
release. Therefore, this commit disables C++ support entirely on the
Blackfin architecture in the internal toolchain backend.

This will avoid a significant number of Blackfin build failures, that
occur when building C++ packages.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>