xserver_xorg-server: move existing patch, and add to 1.17.4
As reported by Manuel in bug #9161, the commit
19b503b0d26fa16fffc980c287ca844077f3f675 that bumped the X.org server
from 1.18.3 to 1.18.4 forgot to also rename the patch directory, making
the existing patch orphan, while it is still needed. This commit
therefore renames the 1.18.3 directory to 1.18.4.

Moreover, this patch is also needed for the 1.17.x X.org server series
which has been added recently, so this commit also adds the patch to a
new 1.17.4 directory.

Fixes bug #9161.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>