wandboard: Use U-Boot 2016.07
Upgrade U-Boot version to 2016.07 and make proper use of U-Boot
distro configuration features.

Provide a boot.scr binary script that is one of the standard ways to
pass a U-Boot script with distro features enabled.

With this method the uboot-env.txt file is no longer needed, so it is
removed.

Signed-off-by: Fabio Estevam <fabio.estevam@nxp.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>