e2fsprogs: fix linking against librt
Our previous patch to link against librt to access clock_gettime() was
merged upstreamed... but it lacked a crucial part which makes the whole
thing useless: $(CLOCK_GETTIME_LIB) is always empty from a make point of
view.

This commit adds a patch to e2fsprogs to fix the problem.

Fixes:

  http://autobuild.buildroot.net/results/41ea60bebfc741604a5499df74363b498ad77a48/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>