docs/manual: document both user-facing br2-external variables
While BR2_EXTERNAL_<NAME>_PATH was already documented, it was not
made obvious that it could be reused in post-biuld, post-image and
in-fakeroot scripts.

The BR2_EXTERNAL_<NAME>_DESC variable was not documented at all.

Update the manual to fix this.

Note: a2x chokes on this:

    .Note:
    Both +BR2_EXTERNAL_$(NAME)_PATH+ and +BR2_EXTERNAL_$(NAME)_DESC+ are
      available [...]

and spits out a totally useless error message:

    a2x: ERROR: "xsltproc" --stringparam toc.section.depth 1 --stringparam
    callout.graphics 0 --stringparam navig.graphics 0 --stringparam
    admon.textlabel 1 --stringparam admon.graphic 0 --stringparam
    chunk.section.depth 0 --output "/home/ymorin/dev/buildroot/O/docs/manual/manual.html"
    "/etc/asciidoc/docbook-xsl/xhtml.xsl"
    "/home/ymorin/dev/buildroot/O/docs/manual/manual.xml" returned
    non-zero exit status 6

So, we had to resort to using different quoting styles for each
variables. They are not semantically equivalent, but the rendering
is the same with the default CSS (which we are using). That's gonna
be good-enough for now...

Reported-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Benoît Allard <benoit.allard@greenbone.net>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>