package/pseudo: new package
Pseudo is a program that can either be used directly or as an
LD_PRELOAD, which allows these operations to succeed as if the
user did have system administrator privileges even though they
are an ordinary user.

Pseudo has a lot of similarities to fakeroot but is a new
implementation that improves on the problems seen using fakeroot.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Andrey Yurovsky <yurovsky@gmail.com>
Cc: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>