package/sane-backends: fix musl build issue
We should include config.h from sanei_backend.h in order to use the
correct if/else HAVE_FOO.

For some reason with Glibc or uClibc there is no problem but with musl
we have the following weird issue:

In file included from epsonds.h:41:0,
                 from epsonds-jpeg.c:18:
../include/sane/sanei_backend.h:99:33: error: expected ';', identifier or '(' before 'int'

../include/sane/sanei_backend.h:99:33: warning: useless type name in empty declaration

That's because HAVE_SIGPROCMASK is not defined although it's correctly
detected by the configure script.

$ grep config.log
config.log:#define HAVE_SIGPROCMASK 1

So, include config.h to avoid to redefine sigset_t.

Fixes:
http://autobuild.buildroot.net/results/9f1/9f1f1cb727b5c5407e69172280a3dee880e55cdf

Signed-off-by: Romain Naour <romain.naour@gmail.com>
Cc: Frank Hunleth <fhunleth@troodon-software.com>
Reviewed-by: Frank Hunleth <fhunleth@troodon-software.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>